+++ cracker-client/src/+components/ApolloWrapper.jsx
@@ -2,7 +2,7 @@ import React from "react";
 import { ApolloClient, InMemoryCache, ApolloProvider } from "@apollo/client";
 import { setContext } from "@apollo/client/link/context";
 import { createUploadLink } from "apollo-upload-client";
-import { useAccessToken } from "./+hooks/useAccessToken";
+import { useAccessToken } from "../+hooks/useAccessToken";
 
 export const ApolloWrapper = ({ children }) => {
   const accessToken = useAccessToken();
